Controller works for fd, bd, home

+++ src/slogo/commands/turtlecommands/Home.java
@@ -17,7 +17,7 @@ public class Home implements TurtleCommand {
     private static final double homeY = 0;
     private TurtleCommand go;
 
-    public Home(int xMax, int yMax, String mode){
+    public Home(double xMax, double yMax, String mode){
         go = new SetPosition(new Constant(homeX), new Constant(homeY), xMax, yMax, mode);
     }
